Refactor FloatingBooks to use SelectedBookView component
